TimeSpan - better manage timespans
